test: fix image article tests
